requirements.txt
fastapi
uvicorn[standard]
ultralytics==8.4.21
opencv-python-headless
pillow
python-multipart
git+https://github.com/ultralytics/CLIP.git